feat: add environment variable support and refactor API core module
- Introduced custom fields in docusaurus.config.ts to pass environment variables to the client-side.
- Refactored the Core.ts module to improve organization by separating environment detection, API discovery, and HTTP client configuration into distinct modules.
- Enhanced the API core module with clear documentation and exports for better maintainability and usability.
